+++ doc/whatsnew.rst
@@ -19,7 +19,8 @@ Version 1.0.3 Release Notes (unreleased)
 
 New features:
 
-Bug fixes:
+Bug fixes/enhancements:
+- do not overwrite user-specified figure titles in Model.plot() functions and allow setting with 'title' keyword argument (PR #711)
 
 Various:
 - update asteval dependency to >=0.9.22 to avoid DeprecationWarnings from NumPy v1.20.0 (PR #707)